+++ CHANGELOG.md
@@ -3,8 +3,51 @@
 ### Changed
 
 ### Added
+
+### Fixed
+* [opensearch role] Possible nil pointer on not setting tenant permission
+
+## [2.0.7] - 2022-12-06
+### Changed
+* Fix id normalization for opensearch 1.1+
+
+
+## [2.0.6] - 2022-10-25
+### Added
+* [index] Add similarity module config in index resource (#321)
+
+
+## [2.0.5] - 2022-10-03
+### Added
+* [index] add ability to configure char_filter on an index (#314)
+
+### Fixed
+* [provider] Return Errors from AWS HTTP Client fixing "Plugin did not respond" (#251)
+
+
+## [2.0.4] - 2022-08-08
+### Changed
+* [ci] Use go1.18 to test and build
+
+
+## [2.0.3] - 2022-08-07
+### Added
+* [index] Add index sorting fields
+
+### Fixed
+* [xpack ilm] Ignore read only element on diff
+* [kibana alert] Handle diff on >=7.11
+* [xpack role] rename should force recreation
+
+
+## [2.0.2] - 2022-06-01
+### Changed
+* Provide context with ping timeout errors and make the timeout configurable
+
+### Added
 * Add custer settings resource
 * Add data stream resource
+* Add opensearch audit config resource
 
 ### Fixed
 * [role mapping] Check error before referencing response